use auth for getting all events
Getting all events is now context aware. You have to have a JWT to hit the endpoint since it uses the JWT to get the user id to use to get events. Only events that the user id from the JWT is an admin of will be returned. If no JWT is passed in, a 401 is returned.